requirements.txt
pandas==2.2.3
streamlit==1.43.2
matplotlib==3.10.1
seaborn==0.13.2
google-cloud-bigquery==3.30.0
db-dtypes==1.4.2
plotly-express==0.4.1
pytest==8.3.5